core: structure / exceptions / visibility
 - some move arounds + make types private that
   should not be exposed to users.

 - add missing es exception grpc conversions.

 - move `Streams` syntax into `Streams` companion
   object.

 - remove type alias in api package.

 - move `Result` into `MetaStreams` companion
   object.
