+++ sec-core/src/main/scala/sec/api/cluster/grpc/Listener.scala
@@ -22,11 +22,11 @@ package grpc
 import io.grpc.NameResolver.{Listener2, ResolutionResult}
 import cats.effect.Sync
 
-trait Listener[F[_]] {
+private[sec] trait Listener[F[_]] {
   def onResult(result: ResolutionResult): F[Unit]
 }
 
-object Listener {
+private[sec] object Listener {
   def apply[F[_]](l2: Listener2)(implicit F: Sync[F]): Listener[F] =
     (rr: ResolutionResult) => F.delay(l2.onResult(rr))
 }
