+++ sec-core/src/main/scala/sec/api/direction.scala
@@ -23,10 +23,11 @@ import cats.Eq
 
 sealed trait Direction
 object Direction {
+
   case object Forwards  extends Direction
   case object Backwards extends Direction
 
-  implicit final class DirectionOps(val d: Direction) extends AnyVal {
+  implicit final private[sec] class DirectionOps(val d: Direction) extends AnyVal {
     def fold[A](fw: => A, bw: => A): A = d match {
       case Forwards  => fw
       case Backwards => bw
