+++ sec-core/src/main/scala/sec/api/grpc/metadata.scala
@@ -21,7 +21,7 @@ package grpc
 import io.grpc.Metadata
 import grpc.constants.Headers.{Authorization, ConnectionName, RequiresLeader}
 
-object metadata {
+private[sec] object metadata {
 
   private[grpc] object keys {
     val authorization: Metadata.Key[UserCredentials] = Metadata.Key.of(Authorization, UserCredentialsMarshaller)
