+++ sec-core/src/main/scala/sec/api/package.scala
@@ -34,9 +34,4 @@ package object api {
 
 //======================================================================================================================
 
-  type DeleteResult = sec.api.Streams.DeleteResult
-  type WriteResult  = sec.api.Streams.WriteResult
-
-//======================================================================================================================
-
 }
